Move parsePncUpdateDataSetXml function into root lib folder
