+++ packages/core/comparison/lib/comparePhase3.ts
@@ -11,10 +11,10 @@ import type ComparisonResultDetail from "../types/ComparisonResultDetail"
 import type { ComparisonResultDebugOutput } from "../types/ComparisonResultDetail"
 
 import CoreAuditLogger from "../../lib/CoreAuditLogger"
+import { parsePncUpdateDataSetXml } from "../../lib/parse/parsePncUpdateDataSetXml"
 import { PncApiError } from "../../lib/PncGateway"
 import serialiseToXml from "../../lib/serialise/pncUpdateDatasetXml/serialiseToXml"
 import getMessageType from "../../phase1/lib/getMessageType"
-import { parsePncUpdateDataSetXml } from "../../phase2/parse/parsePncUpdateDataSetXml"
 import { isPncLockError } from "../../phase3/exceptions/generatePncUpdateExceptionFromMessage"
 import { MAXIMUM_PNC_LOCK_ERROR_RETRIES } from "../../phase3/lib/updatePnc"
 import phase3Handler from "../../phase3/phase3"
